product/requirements.txt
cffi==1.15.0
click==8.1.2
Deprecated==1.2.13
Flask==2.1.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pyparsing==3.0.7
python-dotenv==0.20.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
wrapt==1.14.0